requirements.txt
absl-py==0.8.1
aioconsole==0.3.2
aiohttp>=3.7.4
aiohttp-apispec==2.2.1
aiomonitor==0.4.5
alabaster==0.7.12
alembic==1.4.3
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.3
aniso8601==8.0.0
apispec==3.3.2
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asgiref==3.2.3
asn1crypto==1.2.0
astor==0.8.0
astroid==2.3.3
astropy==3.2.3
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn>=20.12.3
autokeras==1.0.2
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
base58==2.0.1
beautifulsoup4==4.8.1
bidict==0.21.3
bitarray==1.2.2
bitcoinlib==0.6.0
bkcharts==0.2
blake2b-py==0.1.3
blaze==0.11.3
bleach>=3.3.0
blinker==1.4
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
cached-property==1.5.2
cachetools==3.1.1
certifi==2020.12.5
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.6
cheroot==8.4.5
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
cmake==3.18.4.post1
colorama==0.4.1
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography>=3.2
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.8.1
dateparser==0.7.4
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.1
distributed==2.8.1
docutils==0.15.2
dominate==2.5.2
ecdsa==0.17.0
elasticsearch==7.9.1
entrypoints==0.3
et-xmlfile==1.0.1
expiringdict==1.2.1
ez-setup==0.9
fastcache==1.1.0
fastecdsa==2.2.1
filelock==3.0.12
Flask==1.1.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
Flask-Cors>=3.0.9
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fsspec==0.6.0
future==0.18.2
gast==0.3.3
gevent==1.4.0
gevent-websocket==0.10.1
glob2==0.7
gmpy2==2.0.8
greenlet==0.4.15
grpcio==1.25.0
guess-language==0.2
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
hexbytes==0.2.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==0.23
incremental==17.5.0
inflection==0.3.1
intel-openmp==2019.0
ipaddress==1.0.23
ipfshttpclient==0.7.0a1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jaraco.functools==3.0.1
jdcal==1.4.1
jedi==0.15.1
Jinja2>=2.11.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab>=1.2.21
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.6
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
libnacl==1.8.0
lief==0.9.0
lightgbm==2.3.0
llvmlite==0.30.0
locket==0.2.0
lru-dict==1.1.6
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.13.0
matplotlib==3.1.1
mccabe==0.6.1
mglearn==0.1.7
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mlxtend==0.17.2
more-itertools==8.7.0
mpmath==1.1.0
msgpack==0.6.1
multiaddr==0.0.9
multidict==5.1.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==4.4.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.4
nose==1.3.7
notebook>=6.4.1
ntplib==0.4.0
numba==0.46.0
numexpr==2.7.1
numpy==1.19.5
numpydoc==0.9.1
oauthlib==3.1.0
odo==0.5.1
olefile==0.46
opencv-python==4.3.0.36
openpyxl==3.0.1
opt-einsum==3.1.0
p2p-python==3.0.5
packaging==19.2
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.1
partd==1.1.0
path.py==12.0.2
pathlib==1.0.1
pathlib2==2.3.5
patsy==0.5.0
pep8==1.7.1
pexpect==4.7.0
pipenv==2020.8.13
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.12.0
psutil>=5.6.6
psycopg2==2.8.6
ptyprocess==0.6.0
py>=1.10.0
py-ecc==4.1.0
py-evm==0.3.0a19
py-geth==2.4.0
py-solc==3.2.0
py-solc-x==1.0.0
pyaes==1.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycryptodome==3.10.1
pycryptodomex==3.10.1
pycurl==7.43.0.3
pyethash==0.1.27
pyflakes==2.1.1
PyHamcrest==2.0.2
pylint==2.4.4
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyotp==2.6.0
pyp2p==0.8.2
pyparsing==2.4.5
pyroute2==0.6.4
pyroute2.core==0.6.4
pyroute2.ethtool==0.6.4
pyroute2.ipdb==0.6.4
pyroute2.ipset==0.6.4
pyroute2.ndb==0.6.4
pyroute2.nftables==0.6.4
pyroute2.nslink==0.6.4
pyrsistent==0.15.6
PySocks==1.7.1
pystan==2.19.0.0
pystun3==1.0.0
pyTelegramBotAPI==3.6.7
pytest==5.3.0
pytest-arraydiff==0.3
pytest-astropy==0.6.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
python-utils==2.4.0
pytrends==4.4.0
pytz==2019.3
PyWavelets==1.0.3
pyzmq==18.1.0
qtconsole==4.6.0
QtPy==1.9.0
regex==2020.4.4
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rlp==2.0.0a1
rope==0.14.0
rpa==1.27.1
rsa>=4.7
rusty-rlp==0.1.15
scrypt==0.8.18
seaborn==0.9.0
semantic-version==2.8.5
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
snowballstemmer==2.0.0
socketIO-client==0.7.2
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.4.23
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.4
tables==3.4.4
tabulate==0.8.7
tagui==1.27.1
tblib==1.5.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.3
trie==2.0.0a4
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.0.0
ujson==3.2.0
unicodecsv==0.14.1
UPnPy==1.1.6
urllib3>=1.26.5
varint==1.0.2
virtualenv==20.0.34
virtualenv-clone==0.5.4
visitor==0.1.3
wcwidth==0.1.7
webargs==5.5.3
webencodings==0.5.1
websocket-client==0.57.0
websockets>=9.1
Werkzeug==0.16.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.3.3
wurlitzer==2.0.0
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.2.6
xlwings==0.16.0
xlwt==1.3.0
yarl==1.6.3
